--- src/main/java/com/brain/ion/TickCall.java
@@ -1,18 +1,0 @@
-package main.java.com.brain.ion;
-
-/**
- * The interface that allows the main game loop to call the tick method of the
- * game. Add this to the class that should be called when it is time to
- * calculate a game tick.
- * 
- * @author Marshall Brain
- *
- */
-public interface TickCall {
-	
-	/**
-	 * The method that is called when a game tick should be calculated.
-	 */
-	void tick();
-	
-}
