+++ src/main/java/com/brain/ion/settings/SettingEntry.java
@@ -1,0 +1,29 @@
+package main.java.com.brain.ion;
+
+/**
+ * Enum that is used to retrieve a setting from the EnumMap used to store
+ * settings for the game.
+ * 
+ * @author Marshall Brain
+ *
+ */
+public enum SettingEntry {
+	
+	/**
+	 * Key for the initial x coordinate for the main JFrame
+	 */
+	WINDOWED_POS_X,
+	/**
+	 * Key for the initial y coordinate for the main JFrame
+	 */
+	WINDOWED_POS_Y,
+	/**
+	 * Key for the initial width of the main JFrame
+	 */
+	WINDOWED_WIDTH,
+	/**
+	 * Key for the initial height of the main JFrame
+	 */
+	WINDOWED_HEIGHT,
+	
+}
