--- src/resorses/gfx/tool_tips.xml
@@ -1,35 +1,0 @@
-<ion>
-
-	<vector_layer type="body_tool_tip" id="">
-		<rect
-			style="stroke:#000000;stroke-opacity:1;stroke-width:3;
-						fill:#ffffff;fill-opacity:1"
-			cx="0"
-			cy="0"
-			w="1"
-			h="1"
-		/>
-		<text
-			style="fill:#000000;fill-opacity:1;font-size:30;font-family:Open Sans"
-			padding_x="5"
-			padding_y="5"
-		>Error</text>
-	</vector_layer>
-
-	<vector_layer type="body_tool_tip" id="pc_frozen">
-		<rect
-			style="stroke:#000000;stroke-opacity:1;stroke-width:3;
-						fill:#ffffff;fill-opacity:1"
-			cx="0"
-			cy="0"
-			w="1"
-			h="1"
-		/>
-		<text
-			style="fill:#0000ff;fill-opacity:1;font-size:30;font-family:Open Sans"
-			padding_x="5"
-			padding_y="5"
-		>Error</text>
-	</vector_layer>
-	
-</ion>
